Update product.js
## Description

Updating the body of the function findProduct(a, b) for code optimization purpose.

Earlier the issue is that the function stores the result of the multiplication of two variables in the third variables and then it returns the result. This increases the space complexity.

Now I removed the third variable and returned the expression of the multiplication directly.