Update product.js
## Description

Updating the function findProduct(a, b) for code optimization purpose.

The function earlier used to store the values of multiplication of two variables inside a third variable and then returns this value os third variable as answer. It increases the space complexity.

Now It returns directly the express a * b. This expression directly returns the result of multiplication of two variables.