Update product.js
## Description

Updated the function findProduct(a, b) body for the purpose of space optimization.

Earlier the function use to store the value of multiplication inside another variable that takes the extra space and increases the space complexity. To reduce the space complexity I changed the function body and removed the extra variable and it directly return the expression a * b that reduces the space complexity.